22GFLOP/s until 4096

+++ common.h
@@ -36,7 +36,7 @@ void allclose(float *a, float *b, int numel, float rtol) {
 
 void rand_init(float *m, int numel) {
   for (int i = 0; i < numel; i++) {
-    m[i] = (float)rand() / RAND_MAX;
+    m[i] = rand() / RAND_MAX;
   }
 }
